+++ AudioMlSpecTools/wav.py
@@ -2,7 +2,6 @@
 # Global Imports
 ###############################################################################
 from enum import Enum
-import math
 import os
 from typing import Optional
 import warnings
@@ -13,6 +12,11 @@ import warnings
 import soundfile
 import torch
 import torchcodec
+
+###############################################################################
+# Local Imports
+###############################################################################
+from AudioMlSpecTools.resample import resample, Resample
 
 
 ###############################################################################
@@ -36,7 +40,7 @@ class BitsPerSample(Enum):
 ###############################################################################
 # Functions
 ###############################################################################
-def set_audio_length(wave: torch.Tensor, sr: int, duration_secs: Optional[int]):
+def set_audio_length(wave: torch.Tensor, sr: int, duration_secs: Optional[float], pad=False):
     if duration_secs is None:
         return wave
     else:
@@ -46,14 +50,15 @@ def set_audio_length(wave: torch.Tensor, sr: int, duration_secs: Optional[int]):
         st_idx, end_idx = int(start_secs * sr), int(end_secs * sr)
         wave = wave[:, st_idx:end_idx]
 
+        target_num_samples = int(sr * duration_secs)
+
         # Zero Padding
-        num_samples = int(sr * duration_secs)
-        padding_size = max(num_samples - wave.size(1), 0)
+        padding_size = max(target_num_samples - wave.size(1), 0) if pad else 0
         if padding_size > 0:
             wave = torch.cat([wave, torch.zeros(1, padding_size)], dim=1)
 
         # Trim
-        wave = wave[:, :num_samples]
+        wave = wave[:, :target_num_samples]
 
         return wave
 
@@ -62,6 +67,7 @@ def load_wav(path: str,
              *,
              target_sr: Optional[int] = None,
              duration_secs: Optional[int] = None,
+             mono=True,
              ) -> torch.Tensor:
     '''
     Load a WAV file into memory for processing.
@@ -77,26 +83,10 @@ def load_wav(path: str,
     '''
 
     audio = torchcodec.decoders.AudioDecoder(path).get_all_samples()
-    wave, final_sr = _prepare_audio(audio, target_sr)
+    wave, final_sr = _prepare_audio(audio, target_sr=target_sr, mono=mono)
     wave = set_audio_length(wave, final_sr, duration_secs)
 
     return wave
-
-
-def load_wav_as_is(path: str) -> tuple[torch.Tensor, int]:
-    '''
-    Load a WAV file into memory for processing.
-
-    Other file types may work but have not been tested.
-
-    Positional arguments:
-        path -- Location of the audio file
-
-    Returns:
-        Tuple of: entire wav file as a PyTorch Tensor, sample rate
-    '''
-    audio = torchcodec.decoders.AudioDecoder(path).get_all_samples()
-    return _prepare_audio(audio)
 
 
 def is_multichannel(wave: torch.Tensor) -> bool:
@@ -115,44 +105,79 @@ def list_audio_files(dir: str) -> list[str]:
 # Classes
 ###############################################################################
 class WavReader:
-    def __init__(self, filename: str, target_sr: int):
-        self.filename = filename
-        self.target_sr = target_sr
+    def __init__(self,
+                 *,
+                 target_sr: Optional[int] = None,
+                 pad=False,
+                 mono=True,
 
-    def load(self, *, start_sec: Optional[int], end_sec: Optional[int], pad=False) -> torch.Tensor:
-        if start_sec is None and end_sec is None:
-            return load_wav(self.filename, target_sr=self.target_sr)
-        elif start_sec and start_sec < 0:
+                 # For efficient resampling
+                 expected_file_sr: Optional[int] = None,
+                 ):
+        self.target_sr = target_sr
+        self.pad = pad
+        self.mono = mono
+
+        self.expected_file_sr = expected_file_sr
+
+        if self.target_sr and self.expected_file_sr:
+            self.resample = Resample(self.expected_file_sr, self.target_sr)
+        else:
+            self.resample = None
+
+    def __call__(self, filename: str, *, start_sec: Optional[float] = None, end_sec: Optional[float] = None):
+        return self.read(filename, start_sec=start_sec, end_sec=end_sec)
+
+    def read(self, filename: str, *, start_sec: Optional[float] = None, end_sec: Optional[float] = None):
+        if start_sec and start_sec < 0:
             raise ValueError("start_sec must be at least zero")
         elif end_sec and end_sec < 0:
             raise ValueError("end_sec must be at least zero")
         elif start_sec and end_sec and end_sec <= start_sec:
             raise ValueError("end_sec must be strictly higher than start_sec if both are provided")
 
-        audio = torchcodec.decoders.AudioDecoder(self.filename).get_samples_played_in_range(start_sec or 0, end_sec)
-        wave, final_sr = _prepare_audio(audio, self.target_sr)
+        f = torchcodec.decoders.AudioDecoder(filename)
 
-        if start_sec and end_sec and pad:
-            wave = set_audio_length(wave, final_sr, end_sec - start_sec)
+        if start_sec is None and end_sec is None:
+            audio = f.get_all_samples()
+        else:
+            audio = f.get_samples_played_in_range(start_sec or 0, end_sec)
+
+        file_sr = audio.sample_rate
+
+        if self.expected_file_sr and self.expected_file_sr != file_sr:
+            msg = f"File {filename} has sample rate {file_sr} but expected {self.expected_file_sr}"
+            if self.target_sr:
+                raise ValueError(msg)
+            else:
+                warnings.warn(msg)
+
+        wave, final_sr = _prepare_audio(audio, target_sr=self.target_sr, mono=self.mono, resampler=self.resample)
+
+        duration = end_sec - (start_sec or 0) if end_sec else None
+        wave = set_audio_length(wave, final_sr, duration, self.pad)
 
         return wave
-
-    def __call__(self, *, start_sec: Optional[int], end_sec: Optional[int], pad=False) -> torch.Tensor:
-        return self.load(start_sec=start_sec, end_sec=end_sec, pad=pad)
 
 
 ###############################################################################
 # Helpers
 ###############################################################################
-def _prepare_audio(audio: torchcodec.AudioSamples, target_sr: Optional[int] = None):
+def _prepare_audio(audio: torchcodec.AudioSamples,
+                   *,
+                   resampler: Optional[Resample] = None,
+                   target_sr: Optional[int] = None,
+                   mono=True,
+                   ):
     wave = audio.data
 
     # Resample
     final_sr = target_sr or audio.sample_rate
-    wave = resample(wave, orig_freq=audio.sample_rate, new_freq=target_sr)
+    wave = resampler(wave) if resampler else resample(wave, orig_freq=audio.sample_rate, new_freq=target_sr)
 
     # Stereo -> Mono
-    wave = to_mono(wave)
+    if mono:
+        wave = to_mono(wave)
 
     return wave, final_sr
 
@@ -186,136 +211,11 @@ def _prepare_audio(audio: torchcodec.AudioSamples, target_sr: Optional[int] = No
 # CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-###############################################################################
-
-###############################################################################
-# From `torchaudio.functional`
-###############################################################################
-
-def _get_sinc_resample_kernel(
-    orig_freq: int,
-    new_freq: int,
-    gcd: int,
-    lowpass_filter_width: int = 6,
-    rolloff: float = 0.99,
-    device: torch.device = torch.device("cpu"),
-    dtype: Optional[torch.dtype] = None,
-):
-    orig_freq = int(orig_freq) // gcd
-    new_freq = int(new_freq) // gcd
-
-    if lowpass_filter_width <= 0:
-        raise ValueError("Low pass filter width should be positive.")
-    base_freq = min(orig_freq, new_freq)
-    # This will perform antialiasing filtering by removing the highest frequencies.
-    # At first I thought I only needed this when downsampling, but when upsampling
-    # you will get edge artifacts without this, as the edge is equivalent to zero padding,
-    # which will add high freq artifacts.
-    base_freq *= rolloff
-
-    # The key idea of the algorithm is that x(t) can be exactly reconstructed from x[i] (tensor)
-    # using the sinc interpolation formula:
-    #   x(t) = sum_i x[i] sinc(pi * orig_freq * (i / orig_freq - t))
-    # We can then sample the function x(t) with a different sample rate:
-    #    y[j] = x(j / new_freq)
-    # or,
-    #    y[j] = sum_i x[i] sinc(pi * orig_freq * (i / orig_freq - j / new_freq))
-
-    # We see here that y[j] is the convolution of x[i] with a specific filter, for which
-    # we take an FIR approximation, stopping when we see at least `lowpass_filter_width` zeros crossing.
-    # But y[j+1] is going to have a different set of weights and so on, until y[j + new_freq].
-    # Indeed:
-    # y[j + new_freq] = sum_i x[i] sinc(pi * orig_freq * ((i / orig_freq - (j + new_freq) / new_freq))
-    #                 = sum_i x[i] sinc(pi * orig_freq * ((i - orig_freq) / orig_freq - j / new_freq))
-    #                 = sum_i x[i + orig_freq] sinc(pi * orig_freq * (i / orig_freq - j / new_freq))
-    # so y[j+new_freq] uses the same filter as y[j], but on a shifted version of x by `orig_freq`.
-    # This will explain the F.conv1d after, with a stride of orig_freq.
-    width = math.ceil(lowpass_filter_width * orig_freq / base_freq)
-    # If orig_freq is still big after GCD reduction, most filters will be very unbalanced, i.e.,
-    # they will have a lot of almost zero values to the left or to the right...
-    # There is probably a way to evaluate those filters more efficiently, but this is kept for
-    # future work.
-    idx_dtype = dtype if dtype is not None else torch.float64
-
-    idx = torch.arange(-width, width + orig_freq, dtype=idx_dtype, device=device)[None, None] / orig_freq
-
-    t = torch.arange(0, -new_freq, -1, dtype=dtype, device=device)[:, None, None] / new_freq + idx
-    t *= base_freq
-    t = t.clamp_(-lowpass_filter_width, lowpass_filter_width)
-
-    # we do not use built in torch windows here as we need to evaluate the window
-    # at specific positions, not over a regular grid.
-    window = torch.cos(t * math.pi / lowpass_filter_width / 2) ** 2
-
-    t *= math.pi
-
-    scale = base_freq / orig_freq
-    kernels = torch.where(t == 0, torch.tensor(1.0).to(t), t.sin() / t)
-    kernels *= window * scale
-
-    if dtype is None:
-        kernels = kernels.to(dtype=torch.float32)
-
-    return kernels, width
-
-
-def _apply_sinc_resample_kernel(
-    waveform: torch.Tensor,
-    orig_freq: int,
-    new_freq: int,
-    gcd: int,
-    kernel: torch.Tensor,
-    width: int,
-):
-    orig_freq = int(orig_freq) // gcd
-    new_freq = int(new_freq) // gcd
-
-    # pack batch
-    shape = waveform.size()
-    waveform = waveform.view(-1, shape[-1])
-
-    num_wavs, length = waveform.shape
-    waveform = torch.nn.functional.pad(waveform, (width, width + orig_freq))
-    resampled = torch.nn.functional.conv1d(waveform[:, None], kernel, stride=orig_freq)
-    resampled = resampled.transpose(1, 2).reshape(num_wavs, -1)
-    target_length = torch.ceil(torch.as_tensor(new_freq * length / orig_freq)).long()
-    resampled = resampled[..., :target_length]
-
-    # unpack batch
-    resampled = resampled.view(shape[:-1] + resampled.shape[-1:])
-    return resampled
-
-
-def resample(
-    waveform: torch.Tensor,
-    orig_freq: int,
-    new_freq: Optional[int],
-    *,
-    lowpass_filter_width: int = 6,
-    rolloff: float = 0.99,
-) -> torch.Tensor:
-    if orig_freq == new_freq or new_freq is None:
-        return waveform
-
-    gcd = math.gcd(int(orig_freq), int(new_freq))
-
-    kernel, width = _get_sinc_resample_kernel(
-        orig_freq,
-        new_freq,
-        gcd,
-        lowpass_filter_width,
-        rolloff,
-        waveform.device,
-        waveform.dtype,
-    )
-    resampled = _apply_sinc_resample_kernel(waveform, orig_freq, new_freq, gcd, kernel, width)
-    return resampled
-
-
-###############################################################################
+#
 # From `torchaudio._backend.soundfile`
 # Updated by Ryan Quinn 12 June 2026
 ###############################################################################
+
 def _get_subtype(dtype: torch.dtype,
                  encoding: Optional[AudioEncoding],
                  bits_per_sample: Optional[BitsPerSample],